# plus-ms-product/openapi.yaml
openapi: 3.0.3
info:
  title: Plus Gestão — Microsserviço de Produto
  description: |
    API do microsserviço de Produto do sistema Plus Gestão (vestuário plus size).
    Este serviço é responsável pelo **cadastro e gerenciamento do catálogo de produtos**,
    incluindo variantes (itens de grade), tamanhos e associação de mídias.

    ### Dependências externas
    - **MS Categorias** (Grupo 5): fornece `categoriaId` para classificação de produtos
    - **MS Fornecedores** (Grupo It Girls): fornece `fornecedorId` para rastreabilidade

    ### Serviços que nos consomem
    - **MS Estoque** (Grupo 16): usa `itemDeGradeId` para controle de quantidades
    - **MS Pedidos** (Grupo 8): usa produtos e variantes para criação de pedidos
    - **MS Consulta** (Grupo 67): usa endpoints de busca para exibir o catálogo
  version: 1.0.0
  contact:
    name: Grupo 7 — Equipe Produto
  license:
    name: Uso interno — ES2

servers:
  - url: http://localhost:3002
    description: Servidor de desenvolvimento local
  - url: http://plus-ms-product:3002
    description: Servidor Docker (rede interna)

tags:
  - name: Produtos
    description: CRUD e busca de produtos
  - name: Variantes
    description: Itens de grade (combinações de cor + tamanho)
  - name: Tamanhos
    description: Cadastro de tamanhos disponíveis

security:
  - bearerAuth: []

paths:
  /products:
    post:
      tags: [Produtos]
      summary: Criar um novo produto
      description: |
        Cria um produto no catálogo. O produto nasce como `ativo = true` por padrão.
        Aceita um campo opcional `variantes` para criar itens de grade na mesma transação.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
            example:
              nome: "Calça Jeans Skinny"
              descricao: "Calça jeans skinny com elastano para maior conforto"
              marca: "PlusWear"
              preco: 189.90
              categoriaId: "1"
              fornecedorId: "f1e2d3c4-b5a6-7890-abcd-ef1234567890"
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Dados inválidos (ex. categoriaId inexistente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflito de integridade (ex. SKU duplicado em variantes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Erro de validação de campos
    get:
      tags: [Produtos]
      summary: Listar todos os produtos
      description: Retorna uma lista paginada de produtos. Por padrão retorna apenas ativos.
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: ativo
          in: query
          description: Filtrar por status ativo/inativo
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Lista paginada de produtos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductResponse'

  /products/search:
    get:
      tags: [Produtos]
      summary: Buscar produtos com filtros
      description: |
        Busca produtos com filtros combinados (AND lógico). Todos os filtros são opcionais.

        **Exemplos:**
        - `GET /products/search?cor=Azul&tamanho=G` — produtos azuis tamanho G
        - `GET /products/search?nome=jeans&precoMin=100&precoMax=300` — jeans entre R$100 e R$300
      operationId: searchProducts
      parameters:
        - name: nome
          in: query
          description: Busca textual no nome e descrição (case-insensitive)
          schema:
            type: string
        - name: categoriaId
          in: query
          description: Filtrar por ID da categoria (MS Categorias, numérico)
          schema:
            type: string
        - name: fornecedorId
          in: query
          description: Filtrar por ID do fornecedor (MS Fornecedores, UUID)
          schema:
            type: string
        - name: cor
          in: query
          description: Filtrar por cor da variante (case-insensitive)
          schema:
            type: string
        - name: tamanho
          in: query
          description: Filtrar por nome do tamanho (ex. P, M, G, GG)
          schema:
            type: string
        - name: marca
          in: query
          description: Filtrar por marca (case-insensitive)
          schema:
            type: string
        - name: precoMin
          in: query
          description: Preço mínimo
          schema:
            type: number
            format: double
        - name: precoMax
          in: query
          description: Preço máximo
          schema:
            type: number
            format: double
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Resultados da busca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProductResponse'

  /products/{id}:
    get:
      tags: [Produtos]
      summary: Buscar produto por ID
      description: Retorna os dados completos de um produto, incluindo suas variantes aninhadas.
      operationId: getProductById
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetailResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Produtos]
      summary: Atualizar um produto
      description: Atualiza os campos enviados de um produto existente. Campos não enviados são ignorados.
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Produto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/{id}/disable:
    patch:
      tags: [Produtos]
      summary: Desativar um produto (soft delete)
      description: |
        Desativa um produto setando `ativo = false`.
        Todas as variantes do produto também são desativadas automaticamente (cascade).
      operationId: disableProduct
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '200':
          description: Produto desativado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Produto desativado com sucesso"
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/{productId}/variants:
    post:
      tags: [Variantes]
      summary: Criar variante para um produto
      description: |
        Cria um item de grade (variante) associado a um produto.
        Cada variante é uma combinação de cor + tamanho + SKU único.
      operationId: createVariant
      parameters:
        - $ref: '#/components/parameters/ProductIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariantCreateRequest'
            example:
              tamanhoId: "d1e2f3a4-b5c6-7890-abcd-ef1234567890"
              cor: "Azul"
              sku: "CJS-AZL-G"
      responses:
        '201':
          description: Variante criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantResponse'
        '400':
          description: tamanhoId inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: SKU já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Variantes]
      summary: Listar variantes de um produto
      operationId: listVariantsByProduct
      parameters:
        - $ref: '#/components/parameters/ProductIdPathParam'
      responses:
        '200':
          description: Lista de variantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VariantResponse'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /variants/{id}:
    get:
      tags: [Variantes]
      summary: Buscar variante por ID
      operationId: getVariantById
      parameters:
        - $ref: '#/components/parameters/VariantIdParam'
      responses:
        '200':
          description: Variante encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantResponse'
        '404':
          description: Variante não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Variantes]
      summary: Atualizar uma variante
      operationId: updateVariant
      parameters:
        - $ref: '#/components/parameters/VariantIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariantUpdateRequest'
      responses:
        '200':
          description: Variante atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariantResponse'
        '400':
          description: tamanhoId inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Variante não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: SKU já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /variants/{id}/disable:
    patch:
      tags: [Variantes]
      summary: Desativar uma variante
      operationId: disableVariant
      parameters:
        - $ref: '#/components/parameters/VariantIdParam'
      responses:
        '200':
          description: Variante desativada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Variante não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /sizes:
    post:
      tags: [Tamanhos]
      summary: Criar um tamanho
      description: Cria um novo tamanho no cadastro (ex. P, M, G, GG, XGG).
      operationId: createSize
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SizeCreateRequest'
            example:
              nome: "GG"
              descricao: "Tamanho GG — Busto 110-118cm, Cintura 96-104cm"
      responses:
        '201':
          description: Tamanho criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SizeResponse'
        '409':
          description: Tamanho com este nome já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Tamanhos]
      summary: Listar todos os tamanhos
      operationId: listSizes
      parameters:
        - name: ativo
          in: query
          description: Filtrar por status ativo/inativo
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Lista de tamanhos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SizeResponse'

  /sizes/{id}:
    get:
      tags: [Tamanhos]
      summary: Buscar tamanho por ID
      operationId: getSizeById
      parameters:
        - $ref: '#/components/parameters/SizeIdParam'
      responses:
        '200':
          description: Tamanho encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SizeResponse'
        '404':
          description: Tamanho não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Tamanhos]
      summary: Atualizar um tamanho
      operationId: updateSize
      parameters:
        - $ref: '#/components/parameters/SizeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SizeUpdateRequest'
      responses:
        '200':
          description: Tamanho atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SizeResponse'
        '404':
          description: Tamanho não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /sizes/{id}/disable:
    patch:
      tags: [Tamanhos]
      summary: Desativar um tamanho
      operationId: disableSize
      parameters:
        - $ref: '#/components/parameters/SizeIdParam'
      responses:
        '200':
          description: Tamanho desativado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Tamanho não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT emitido pelo MS Auth (HS256).
        Claims esperadas: `sub` (email), `user_id`, `role` (admin | vendedor).

  parameters:
    ProductIdParam:
      name: id
      in: path
      required: true
      description: ID do produto (UUID)
      schema:
        type: string
        format: uuid
    ProductIdPathParam:
      name: productId
      in: path
      required: true
      description: ID do produto (UUID)
      schema:
        type: string
        format: uuid
    VariantIdParam:
      name: id
      in: path
      required: true
      description: ID da variante (UUID)
      schema:
        type: string
        format: uuid
    SizeIdParam:
      name: id
      in: path
      required: true
      description: ID do tamanho (UUID)
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      description: Número da página (começa em 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Quantidade de itens por página
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  schemas:
    ProductCreateRequest:
      type: object
      required: [nome, preco]
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 255
          example: "Calça Jeans Skinny"
        descricao:
          type: string
          maxLength: 2000
          example: "Calça jeans skinny com elastano"
        marca:
          type: string
          maxLength: 100
          example: "PlusWear"
        preco:
          type: number
          format: double
          minimum: 0
          example: 189.90
        categoriaId:
          type: string
          nullable: true
          description: ID da categoria (MS Categorias, numérico)
          example: "1"
        fornecedorId:
          type: string
          nullable: true
          description: ID do fornecedor (MS Fornecedores, UUID)
          example: "f1e2d3c4-b5a6-7890-abcd-ef1234567890"
        variantes:
          type: array
          description: Lista opcional de variantes a criar junto com o produto
          items:
            $ref: '#/components/schemas/VariantCreateNestedRequest'

    ProductUpdateRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 255
        descricao:
          type: string
          maxLength: 2000
        marca:
          type: string
          maxLength: 100
        preco:
          type: number
          format: double
          minimum: 0
        categoriaId:
          type: string
          nullable: true
        fornecedorId:
          type: string
          nullable: true

    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        descricao:
          type: string
          nullable: true
        marca:
          type: string
          nullable: true
        preco:
          type: number
          format: double
        ativo:
          type: boolean
        categoriaId:
          type: string
          nullable: true
        fornecedorId:
          type: string
          nullable: true
        criadoEm:
          type: string
          format: date-time
        atualizadoEm:
          type: string
          format: date-time

    ProductDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ProductResponse'
        - type: object
          properties:
            variantes:
              type: array
              items:
                $ref: '#/components/schemas/VariantResponse'

    PaginatedProductResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8

    VariantCreateNestedRequest:
      type: object
      required: [tamanhoId, cor, sku]
      properties:
        tamanhoId:
          type: string
          format: uuid
        cor:
          type: string
          maxLength: 50
          example: "Azul"
        sku:
          type: string
          maxLength: 50
          example: "CJS-AZL-G"

    VariantCreateRequest:
      type: object
      required: [tamanhoId, cor, sku]
      properties:
        tamanhoId:
          type: string
          format: uuid
          example: "d1e2f3a4-b5c6-7890-abcd-ef1234567890"
        cor:
          type: string
          maxLength: 50
          example: "Azul"
        sku:
          type: string
          maxLength: 50
          example: "CJS-AZL-G"

    VariantUpdateRequest:
      type: object
      properties:
        tamanhoId:
          type: string
          format: uuid
        cor:
          type: string
          maxLength: 50
        sku:
          type: string
          maxLength: 50

    VariantResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        produtoId:
          type: string
          format: uuid
        tamanhoId:
          type: string
          format: uuid
        cor:
          type: string
        sku:
          type: string
        ativo:
          type: boolean
        tamanho:
          $ref: '#/components/schemas/SizeResponse'

    SizeCreateRequest:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          maxLength: 10
          example: "GG"
        descricao:
          type: string
          maxLength: 500
          example: "Tamanho GG — Busto 110-118cm, Cintura 96-104cm"

    SizeUpdateRequest:
      type: object
      properties:
        nome:
          type: string
          maxLength: 10
        descricao:
          type: string
          maxLength: 500

    SizeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        descricao:
          type: string
          nullable: true
        ativo:
          type: boolean

    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operação realizada com sucesso"

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Recurso não encontrado"